add gitflow and fix fedora vlc and vim-bootstrap

+++ sh/common/vim-bootstrap/main.sh
@@ -1,9 +1,17 @@
 #!/bin/sh
 
-echo -e "${bakcyn}[common][vim-bootstrap] Start Install ${txtrst}";
+if [[ -e /home/$DOTFILE_DEFAULT_USER/.vimrc ]]; then
+    
+    echo -e "${bakgrn}[installed][vim-bootstrap]${txtrst} already installed ;)" ;
+    
+else
 
-mv sh/common/vim-bootstrap/generate.vim /home/$DOTFILE_DEFAULT_USER/.vimrc;
-chmod $DOTFILE_DEFAULT_USER:$DOTFILE_DEFAULT_USER /home/$DOTFILE_DEFAULT_USER/.vimrc;
-su - $DOTFILE_DEFAULT_USER -c "vim +NeoBundleInstall +qall";
+    echo -e "${bakcyn}[common][vim-bootstrap] Start Install ${txtrst}";
 
-echo -e "${bakcyn}[common][vim-bootstrap] Finish Install ${txtrst}";
+    mv sh/common/vim-bootstrap/generate.vim /home/$DOTFILE_DEFAULT_USER/.vimrc;
+    chmod $DOTFILE_DEFAULT_USER:$DOTFILE_DEFAULT_USER /home/$DOTFILE_DEFAULT_USER/.vimrc;
+    su - $DOTFILE_DEFAULT_USER -c "vim +NeoBundleInstall +qall";
+
+    echo -e "${bakcyn}[common][vim-bootstrap] Finish Install ${txtrst}";
+
+fi
